backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
psycopg2-binary==2.9.9
python-dotenv==0.19.0
werkzeug==2.0.1
sqlalchemy==1.4.46
flask-migrate==3.1.0